Refactor : main을 제외한 파일들 test, customFaker, model 폴더로 정리(#1)
기존에 faker관련 파일들과 test관련 파일들이 메인 repo 바로 아래에 너저분하게 있어서 정리해줌.
근데 python 모듈에 대해 잘 몰라서 일단 class화 시켜서 리팩토링 진행.
